+++ src/app/issue/timer-button/timer-button.component.html
@@ -2,7 +2,7 @@
   nz-button
   nzType="primary"
   (click)="handleButtonClicked($event)"
-  [disabled]="disabled"
+  [disabled]="cannotStartTime()"
 >
   <i nz-icon nzType="play-circle" nzTheme="outline" *ngIf="!timerRunning"></i>
   <i nz-icon nzType="pause-circle" nzTheme="outline" *ngIf="timerRunning"></i>
